// YourProjectName/src/Web/ClientApp/src/api/nswag-api.g.ts
export interface BaseEntity {
    id?: number;
    domainEvents?: BaseEvent[];
}

export interface User extends BaseEntity {
    username: string;
    role: string;
}

export interface BaseEvent {
}

export interface TokenModel {
    value?: string;
    expiration?: Date;
}